feat: Creation of the anime pagination route, in addition to making some adjustments to the project

+++ src/useCases/Anime/Delete/Delete.ts
@@ -1,9 +1,25 @@
+import { StorageFirebase } from "../../../providers/IStorageFirebase.ts";
 import { IAnimeRepository } from "../../../repositories/IAnimeRepository.ts";
 import { IAnimesRequestDTO } from "./ListOneDTO.ts";
 
 export class DeleteAnimeUseCase {
   constructor(private animeRepository: IAnimeRepository) {}
   async execute(data: IAnimesRequestDTO) {
-    await this.animeRepository.delete(data.id);
+    try {
+      const dataAnime = await this.animeRepository.listOne(data.id);
+
+      if (dataAnime === null) {
+        throw new Error("Anime não encontrado");
+      }
+
+      if (dataAnime.urlImg) {
+        await StorageFirebase.deleteImg(dataAnime.urlImg, "animes/");
+      }
+
+      await this.animeRepository.delete(data.id);
+
+    } catch (err: any) {
+      throw new Error(err.message);
+    }
   }
 }
